+++ core-utils/src/main/java/com/agileburo/anytype/core_utils/text/EnterKeyDetector.kt
@@ -5,15 +5,16 @@ import android.text.Spanned
 
 class DefaultEnterKeyDetector(
     private val onEndLineEnterClicked: () -> Unit,
-    private val onSplitLineEnterClicked: () -> Unit
+    private val onSplitLineEnterClicked: (Int) -> Unit
 ) : EnterKeyDetector() {
+
     override fun onEndEnterPress(textBeforeEnter: Spanned): CharSequence? {
         onEndLineEnterClicked()
         return EMPTY_REPLACEMENT
     }
 
-    override fun onSplitEnterPress(textBeforeEnter: Spanned): CharSequence? {
-        onSplitLineEnterClicked()
+    override fun onSplitEnterPress(textBeforeEnter: Spanned, index: Int): CharSequence? {
+        onSplitLineEnterClicked(index)
         return EMPTY_REPLACEMENT
     }
 }
@@ -34,14 +35,14 @@ abstract class EnterKeyDetector : InputFilter {
                 if (dend - 1 == dest.lastIndex)
                     onEndEnterPress(textBeforeEnter = dest)
                 else
-                    onSplitEnterPress(textBeforeEnter = dest)
+                    onSplitEnterPress(textBeforeEnter = dest, index = dend)
             }
             else -> null
         }
     }
 
     abstract fun onEndEnterPress(textBeforeEnter: Spanned): CharSequence?
-    abstract fun onSplitEnterPress(textBeforeEnter: Spanned): CharSequence?
+    abstract fun onSplitEnterPress(textBeforeEnter: Spanned, index: Int): CharSequence?
 
     companion object {
         const val EMPTY_REPLACEMENT = ""