+++ data/src/main/java/com/agileburo/anytype/data/auth/repo/block/BlockDataRepository.kt
@@ -63,4 +63,8 @@ class BlockDataRepository(
     override suspend fun merge(command: Command.Merge) {
         factory.remote.merge(command.toEntity())
     }
+
+    override suspend fun split(command: Command.Split) {
+        factory.remote.split(command.toEntity())
+    }
 }